refactor(turn-runner): move the await-edge resolver tests into their own file
